Another big refactor, causing most of the functionality to work again + some extras. Full changelog:
- FPS Counter in top left corner, at all times. You can turn it off by commenting the code out in the main class.
- Other stations work as well, there wasn't any functionality in them though.
- You can now fly around if you enter the navigation station.
- Two games can connect and if the server starts the game, the connected client starts as well.
- FPS is now steady at 60 FPS (on my computer anyway). There was a huge performance problem with the DistributeReceivedPackets function.

Problems:
- The memory heap gets corrupted by the Enemy sync mechanic of the SendAndReceivePackets class. The problem lays in the deletion of Enemy classes. I did not figure this out yet. The compressing and uncompressing of packets with a Enemy vector should be redesigned probably.

+++ source/Game/MainMenuEventReceiver.cpp
@@ -39,7 +39,7 @@ bool MainMenuEventReceiver::OnEvent(const SEvent& event)
 				ipadress[wcslen(inputwchar)] = 0;
 
 				portInput = context.u_interface->getElementWithId(PORT_INPUT_BOX)->getText();
-				port = atof(portInput.c_str());
+				port = atoi(portInput.c_str());
 
 				namewchar = (wchar_t*)context.u_interface->getElementWithId(NAME_INPUT_BOX)->getText();
 				playername = (char*)malloc(wcslen(namewchar)+ 1);
